ugsra/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-active-link==0.1.8
django-allauth==0.56.1
django-ckeditor==6.5.1
django-cors-headers==4.3.0
django-js-asset==2.0.0
filelock==3.9.0
gunicorn==20.1.0
idna==3.4
include==0.2.2
loader==2017.9.11
oauthlib==3.2.2
path==16.6.0
Pillow==9.5.0
platformdirs==3.0.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
urllib3==2.0.4
virtualenv==20.19.0
